Refactor: 파일구조 변경 시 빠진 파일 추가
