+++ src/main/java/com/example/Tokkit_server/notification/service/NotificationSettingService.java
@@ -1,6 +1,6 @@
 package com.example.Tokkit_server.notification.service;
 
-import com.example.Tokkit_server.notification.dto.NotificationCategoryUpdateRequestDto;
+import com.example.Tokkit_server.notification.dto.request.NotificationCategoryUpdateRequestDto;
 import com.example.Tokkit_server.user.dto.response.NotificationCategorySettingResponseDto;
 
 import java.util.List;
